Complete functionality
There are still a couple of bugs. Firstly the text does not update from 'files' to 'lines' and I think this is because the script is actually crashing when calling .increment() as it's not recognised as a method.

The second bug is that it only seems to run on a page refresh, rather than clicking to the 'files' page from another area of the PR.